m68k: added MAXI030 platform
This is very WIP.

100Hz timer (HZ) is generated by the FPGA and uses autovector #1

Serial ports are provided, in hardware, by a SC26C94, but the sccnxp
driver is used in Linux. This driver supports the SC26C92, which is
half of a '94. Am not sure if it is a m68k limitation, but I could
not get threaded interrupts to work so old style ints are used.
Oteherwise this driver is unchanged. Autovector #2.

IDE uses the platform_pata driver. 32 bit instructions are used for
the transfers. Interrupts are not currently used, as I had probs
with "nobody cared" errors. Needs investigation. Autovector #4
will be used.

The onboard RTL8019AS uses a modifed version of the Zorro driver for
this IC. Currently this is limited to 8 bit transfers. Autovector #4
is used. This contains code for 16 bit mode which will hopefully
be resurected.

Currently investigation problems with head.S where the MMU start
causes a lockup, hence the extra debug here.