m68k: maxi030: added framebuffer driver
Framebuffer driver supports the 640x480 8bpp paletted mode
currently implemented in the Cyclone 2 based video card installed
in the first slot. There's even a tux! Shame there is as yet no
keyboard.

Also added a go at a GPIO driver for controlling the FPGA's LED
though am unsure if this shouldn't be a sipmle platform mem based
component.

Renumbered the slots and IO to be at 0x4.. instead of 0x8.. to get
the IO out of the way of the kernels 0x8.. virtual address for RAM.
Whilst it could probably be renumbered with the 030's MMU, this way
doesn't (I think) incure any cycle penalty.